+++ example-custom-colors/android/app/src/main/kotlin/com/example/system_design_custom_colors/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.system_design_custom_colors
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()